third_party/blink/web_tests/platform/win11-arm64/http/tests/images/drag-drop-image-filename-expected.txt

CONSOLE MESSAGE: FAIL: test image failed to load.
